+++ interfaces/src/main/java/ar/edu/itba/paw/service/ReservationService.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 
 public interface ReservationService {
 
-    public List<Reservation> findByUser(int userId);
+    public List<Reservation> findByUser(long userId);
 
-    public List<Reservation> findByRestaurant(int restaurantId);
+    public List<Reservation> findByRestaurant(long restaurantId);
 
     public Optional<Reservation> findById(int id);
